02.02.2026

+++ src/OOP/Corporartion/ShoeCard.kt
@@ -4,7 +4,7 @@ class ShoeCard(
     name: String,
     brand: String,
     price: Int,
-    val size: Int
+    val size: Float
 ): ProductCard(name = name, brand = brand, price = price) {
 
     override fun printInfo() {
